version 0.3

+++ src/main/java/ru/kata/spring/boot_security/demo/models/User.java
@@ -13,13 +13,13 @@ import java.util.Set;
 public class User implements UserDetails {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column
     private String name;
 
     @Column
-    private int age;
+    private Integer age;
 
     private String username;
     private String password;
@@ -46,11 +46,11 @@ public class User implements UserDetails {
         this.roles = roles;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
